# backend/src/config/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Sentinel Security Scanner",
    "description": "Backend API documentation for the API Sentinel AI-powered security scanner. Provides JWT authentication, scan target orchestration, real-time results retrieval, AI-enriched explanations, and PDF report downloads.",
    "version": "1.0.0",
    "contact": {
      "name": "API Sentinel Team",
      "email": "support@sentinel.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/scan": {
      "post": {
        "summary": "Configure a New Scan Target",
        "description": "Creates a new scan configuration in the database for the given target URL.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "targetUrl"
                ],
                "properties": {
                  "targetUrl": {
                    "type": "string",
                    "format": "uri",
                    "example": "http://localhost:8000/test-vulnerabilities"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Scan configured successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Scan created successfully"
                    },
                    "scan": {
                      "$ref": "#/components/schemas/Scan"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid URL target or validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/scans": {
      "get": {
        "summary": "Retrieve User Scans and Dashboard Stats",
        "description": "Returns all scan targets configured by the logged-in user, along with dashboard statistics.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Scans and stats retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scans": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Scan"
                      }
                    },
                    "stats": {
                      "$ref": "#/components/schemas/DashboardStats"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/scan/{id}": {
      "get": {
        "summary": "Retrieve Scan Details & Vulnerability Findings",
        "description": "Returns the details of a single scan config along with all detected vulnerabilities and AI security recommendations.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the scan",
            "schema": {
              "type": "string",
              "example": "651a2e4125fdab31cc8d41e2"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scan details and findings retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scan": {
                      "$ref": "#/components/schemas/Scan"
                    },
                    "findings": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Vulnerability"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Scan not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a Scan Configuration",
        "description": "Deletes the scan configuration and all its associated vulnerability findings and PDF records.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the scan to delete",
            "schema": {
              "type": "string",
              "example": "651a2e4125fdab31cc8d41e2"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scan deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Scan and associated data deleted successfully"
                    },
                    "scanId": {
                      "type": "string",
                      "example": "651a2e4125fdab31cc8d41e2"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Scan": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "651a2e4125fdab31cc8d41e2"
          },
          "targetUrl": {
            "type": "string",
            "example": "http://localhost:8000/test-vulnerabilities"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "SCANNING",
              "COMPLETED",
              "FAILED"
            ],
            "example": "COMPLETED"
          },
          "score": {
            "type": "number",
            "example": 85
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-26T09:30:50.000Z"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-26T09:32:10.000Z"
          },
          "durationMs": {
            "type": "number",
            "example": 1540
          },
          "totalEndpointsScanned": {
            "type": "number",
            "example": 4
          }
        }
      },
      "DashboardStats": {
        "type": "object",
        "properties": {
          "totalScans": {
            "type": "number",
            "example": 3
          },
          "averageScore": {
            "type": "number",
            "example": 92
          },
          "criticalFindings": {
            "type": "number",
            "example": 1
          },
          "lastScanTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-06-26T09:32:10.000Z"
          }
        }
      },
      "Vulnerability": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "651a2e4125fdab31cc8d41f5"
          },
          "endpoint": {
            "type": "string",
            "example": "/test-vulnerabilities"
          },
          "method": {
            "type": "string",
            "example": "GET"
          },
          "issue": {
            "type": "string",
            "example": "Missing Security Headers"
          },
          "severity": {
            "type": "string",
            "enum": [
              "CRITICAL",
              "HIGH",
              "MEDIUM",
              "LOW"
            ],
            "example": "LOW"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "HIGH",
              "MEDIUM",
              "LOW"
            ],
            "example": "HIGH"
          },
          "description": {
            "type": "string",
            "example": "The HTTP response does not include standard security-oriented headers."
          },
          "recommendation": {
            "type": "string",
            "example": "Configure the web server or application framework to send X-Frame-Options and Content-Security-Policy headers."
          },
          "aiExplanation": {
            "type": "string",
            "example": "The application is missing headers like X-Frame-Options, exposing it to clickjacking attacks."
          },
          "aiImpact": {
            "type": "string",
            "example": "Attackers can embed your site inside an iframe to hijack user clicks."
          },
          "aiFix": {
            "type": "string",
            "example": "Use Helmet middleware to apply secure headers in Express.js."
          },
          "codeExample": {
            "type": "string",
            "example": "const helmet = require('helmet');\napp.use(helmet());"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Invalid credentials provided."
          },
          "errorCode": {
            "type": "string",
            "example": "UNAUTHORIZED"
          }
        }
      }
    }
  }
}